Separo el formulario de la pantalla de creeación de evento
