+++ src/components/EventForm/EventFormStyles.css
@@ -1,0 +1,5 @@
+.formContainer {
+  border: 1px solid black;
+  border-radius: 10px;
+  padding: 10px 10px;
+}
