+++ src/screens/CreateEventScreen/CreateEventScreenStyles.css
@@ -4,9 +4,3 @@
   gap: 24px;
   margin: 0 auto;
 }
-
-.formContainer {
-  border: 1px solid black;
-  border-radius: 10px;
-  padding: 10px 10px;
-}
